% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_map.R
\name{shot_map}
\alias{shot_map}
\title{Plot shots from a football match}
\usage{
shot_map(match_id, title = "")
}
\arguments{
\item{match_id}{Desired match ID from understat.com}

\item{title}{Plot title; empty by default}
}
\value{
Interactive ggiraph plot displaying both teams shots side by side printed to the Viewer.
}
\description{
Plot shots from a football match
}
\examples{
shot_map(26631, title = "xG Map")

}
