% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_chart.R
\name{xg_chart}
\alias{xg_chart}
\title{Plot an xG chart of a football match}
\usage{
xg_chart(
  match_id,
  home_team_color,
  away_team_color,
  competition = "",
  bg_color = "#FFF1E5",
  plot_bg_color = "#FFF1E5"
)
}
\arguments{
\item{match_id}{Match ID from understat.com}

\item{home_team_color}{Color used for the home team}

\item{away_team_color}{Color used for the away team}

\item{competition}{Competition name as a subtitle; empty by default.}

\item{bg_color}{Chart background color; defaults to "#FFF1E5"}

\item{plot_bg_color}{Plot background color; defaults to "#FFF1E5"}
}
\value{
Interactive highcharter plot displaying the xG chart of both teams.
}
\description{
Plot an xG chart of a football match
}
\examples{
\dontrun{
xg_chart(26631, "red", "grey", competition = "Premier League")
}
}
