% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_map.R
\name{xg_map}
\alias{xg_map}
\title{Plot shots xG of a football match}
\usage{
xg_map(match_id, title = "")
}
\arguments{
\item{match_id}{Desired match ID from understat.com}

\item{title}{Plot title; empty by default}
}
\value{
Interactive ggiraph transparent plot displaying both teams shots side by side printed to the Viewer.
}
\description{
Plot shots xG of a football match
}
\examples{
\dontrun{
xg_map(26631, title = "xG Map")
}
}
