% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_chart.R
\name{xg_chart}
\alias{xg_chart}
\title{xG chart plotting}
\usage{
xg_chart(
  match_id,
  home_team_color,
  away_team_color,
  competition = "Premier League",
  background_color = "#FFF1E5"
)
}
\arguments{
\item{match_id}{Match ID from understat.com}

\item{home_team_color}{Color used for the home team}

\item{away_team_color}{Color used for the away team}

\item{competition}{Competition name as a subtitle; defaults to "Premier League"}

\item{background_color}{Chart background color; defaults to "#FFF1E5"}
}
\value{
Interactive highcharter plot displaying the xG chart of both teams.
}
\description{
xG chart plotting
}
\examples{
xg_chart(26631, "red", "grey", "#FFF1E5")

}
